api/requirements.txt
flask>=2.0.1
python-dotenv>=0.19.0
PyMuPDF>=1.19.2
pdf2image>=1.16.0
pytesseract>=0.3.8
python-docx>=0.8.11
openai>=1.0.0